changed service time from int to float to coincide with project requirements

+++ process_producer.h
@@ -3,7 +3,7 @@
 
 struct process{
 	int arrival_time;
-	int service_time;
+	float service_time;
 	int priority;
 	char name;
 };
